requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
openai
python-dotenv
PyGithub
tenacity
slowapi
async-lru
pytest
httpx
